+++ Source/General/Project.vb
@@ -50,7 +50,6 @@ Public Class Project
     Public DemuxAudio As DemuxMode = DemuxMode.All
     Public DemuxChapters As Boolean = True
     Public DemuxVideo As Boolean = False
-    Public DoviMode As DoviMode = DoviMode.Mode2
     Public ExtractHdrmetadata As HdrmetadataMode = HdrmetadataMode.All
     Public ExtractTimestamps As TimestampsMode = TimestampsMode.VfrOnly
     Public ExtractForcedSubSubtitles As Boolean = True
@@ -61,7 +60,9 @@ Public Class Project
     Public ForcedOutputModOnlyIfCropped As Boolean = False
     Public HardcodedSubtitle As Boolean
     Public Hdr10PlusMetadataFile As String
-    Public HdrDolbyVisionMetadataFile As String
+    Public HdrDolbyVisionMetadataFiles As IEnumerable(Of DolbyVisionMetadataFile)
+    Public HdrDolbyVisionCropMode As DoviCropMode = DoviCropMode.Cropped
+    Public HdrDolbyVisionMode As DoviMode = DoviMode.Mode2
     Public ImportVUIMetadata As Boolean = True
     Public ITU As Boolean
     Public LastOriginalSourceFile As String